pyproject.toml
[build-system]
requires = [
    "setuptools >= 64",
    "setuptools_scm >= 7",
]
build-backend = "setuptools.build_meta"

[project]
name = "phreeqcinwt"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    "numpy",
    "pyyaml",
    "molmass",
    "phreeqpy",
]